advanced-autonomous-team/requirements.txt
pyautogen>=0.2.35
ollama
langchain
langchain-community
markdown
rich
termcolor
pydantic
